tools/nut-scanner/scan_*.c: when logging "pass" (bool_t), err on the safe side to cast it as (unsigned int) [#3673]
...This may get revisited after we consistently use
one boolean type in the code base.

+++ tools/nut-scanner/scan_ipmi.c
@@ -883,7 +883,7 @@ nutscan_device_t * nutscan_scan_ip_range_ipmi(nutscan_ip_range_list_t * irl, nut
 					" pass=%u",
 					__func__, max_threads_scantype,
 					curr_threads, thread_count,
-					pass
+					(unsigned int)pass
 				);
 			}
 # else
